Fix implicit conversion caught by ubsan
src/models/snes/tileset.cpp:66:26: runtime error: implicit conversion from type 'int' of value 384 (32-bit, signed) to type 'uint_fast8_t' (aka 'unsigned char') changed the value to 128 (8-bit, unsigned)

+++ src/models/snes/tileset.cpp
@@ -56,8 +56,8 @@ static void readSnesTile8px(Tile<8>& tile, const uint8_t* data, const unsigned b
 
             const unsigned biEnd = (b < bitDepth - 1) ? 2 : 1;
             for (unsigned bi = 0; bi < biEnd; bi++) {
-                uint_fast8_t byte = *dataPos++;
-                uint_fast8_t bits = 1 << (b + bi);
+                unsigned byte = *dataPos++;
+                unsigned bits = 1 << (b + bi);
 
                 for (unsigned x = 0; x < TILE_SIZE; x++) {
                     if (byte & 0x80) {
